refactor(convert bytes): Rename for consistency PE-126
